requirements.txt
black==21.5b0
coverage==5.5
Django==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
flake8==3.9.1
flake8-variables-names==0.0.4
isort==5.8.0
